test: reorg
